Implement server-side caching

+++ src/services/postgres/notes-service.js
@@ -42,8 +42,9 @@ function mapDBToNote(data) {
 export class NotesService {
 	/**
 	 * @param {import("./collaborations-service.js").CollaborationsService} collaborationsService
+	 * @param {import("../redis/cache-service.js").CacheService} cacheService
 	 */
-	constructor(collaborationsService) {
+	constructor(collaborationsService, cacheService) {
 		/**
 		 * @private
 		 */
@@ -53,6 +54,11 @@ export class NotesService {
 		 * @private
 		 */
 		this._collaborationsService = collaborationsService;
+
+		/**
+		 * @private
+		 */
+		this._cacheService = cacheService;
 	}
 
 	/**
@@ -76,6 +82,8 @@ export class NotesService {
 			throw new InvariantError("Catatan gagal ditambahkan");
 		}
 
+		await this._cacheService.delete(`notes:${owner}`);
+
 		return result.rows[0].id;
 	}
 
@@ -85,17 +93,25 @@ export class NotesService {
 	 * @return {Promise<Array<Omit<Note, "owner">>>}
 	 */
 	async getNotes(owner) {
-		const result = await this._pool.query({
-			text: `
-        SELECT notes.* FROM notes
-        LEFT JOIN collaborations ON collaborations.note_id = notes.id
-        WHERE notes.owner = $1 OR collaborations.user_id = $1
-        GROUP BY notes.id
-      `,
-			values: [owner],
-		});
+		try {
+			const result = await this._cacheService.get(`notes:${owner}`);
+			return JSON.parse(result);
+		} catch {
+			const result = await this._pool.query({
+				text: `
+          SELECT notes.* FROM notes
+          LEFT JOIN collaborations ON collaborations.note_id = notes.id
+          WHERE notes.owner = $1 OR collaborations.user_id = $1
+          GROUP BY notes.id
+        `,
+				values: [owner],
+			});
+			const mappedResult = result.rows.map(mapDBToNote);
 
-		return result.rows.map(mapDBToNote);
+			await this._cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult));
+
+			return mappedResult;
+		}
 	}
 
 	/**
@@ -132,7 +148,7 @@ export class NotesService {
         UPDATE notes
         SET title = $1, body = $2, tags = $3, updated_at = $4
         WHERE id = $5
-        RETURNING id
+        RETURNING id, owner
       `,
 			values: [title, body, tags, updatedAt, id],
 		};
@@ -142,6 +158,10 @@ export class NotesService {
 		if (!result.rows.length) {
 			throw new NotFoundError("Gagal memperbarui catatan. Id tidak ditemukan");
 		}
+
+		const { owner } = result.rows[0];
+
+		await this._cacheService.delete(`notes:${owner}`);
 	}
 
 	/**
@@ -149,7 +169,7 @@ export class NotesService {
 	 */
 	async deleteNoteById(id) {
 		const query = {
-			text: "DELETE FROM notes WHERE id = $1 RETURNING id",
+			text: "DELETE FROM notes WHERE id = $1 RETURNING id, owner",
 			values: [id],
 		};
 
@@ -158,6 +178,10 @@ export class NotesService {
 		if (!result.rows.length) {
 			throw new NotFoundError("Catatan gagal dihapus. Id tidak ditemukan");
 		}
+
+		const { owner } = result.rows[0];
+
+		await this._cacheService.delete(`notes:${owner}`);
 	}
 
 	/**
